+++ graphlearn/minor/autoabstract/abstractor.py
@@ -1,22 +1,10 @@
-from eden.modifier.graph.structure import contraction
-from eden.graph import Vectorizer
 import networkx as nx
-from graphlearn.utils import draw
-import eden
-import multiprocessing as mp
-from GArDen.decompose import ThresholdedConnectedComponents
-import copy
 from GArDen.compose import Flatten
+from GArDen.decompose import ThresholdedConnectedComponents
+from eden.graph import Vectorizer
+from eden.modifier.graph.structure import contraction
 
-
-def node_operation(graph, f):
-    # applies function to n,d of nodes(data=True)
-    # if you want to do assignments do this: def my(n,d): d[dasd]=asd
-    # if you want the result you may use lambda :)
-    res = []
-    for n, d in graph.nodes(data=True):
-        res.append(f(n, d))
-    return res
+from graphlearn.utils import node_operation, remove_eden_annotation
 
 
 class GraphToAbstractTransformer(object):
@@ -26,8 +14,15 @@ class GraphToAbstractTransformer(object):
     this class is just a helper for minor transform.
     '''
 
-    def __init__(self, estimator=None, score_threshold=0.0, min_size=0, max_size=50, debug=False, layer=False,
-                 vectorizer=Vectorizer()):
+    def __init__(self, estimator=None,
+                 score_threshold=0.0,
+                 min_size=0,
+                 max_size=50,
+                 debug=False,
+                 layer=False,
+                 vectorizer=Vectorizer(),
+                 score_attribute='importance',
+                 group_attribute ='class'):
         '''
         Parameters
         ----------
@@ -51,8 +46,10 @@ class GraphToAbstractTransformer(object):
         self.debug = debug
         self.max_size = max_size
         self.layer = layer
+        self.score_attribute=score_attribute
+        self.group_attribute = group_attribute
 
-    def get_subgraphs(self, inputs, score_attribute='importance', group='class'):
+    def get_subgraphs(self, inputs):
         '''
         calls get_subgraph_single on all graphs, has option to go multiprocess
 
@@ -69,10 +66,14 @@ class GraphToAbstractTransformer(object):
             yields subgraphs
         '''
 
-        return standalone_get_subgraphs(inputs, score_attribute, group, self.score_threshold, self.min_size,
+        return standalone_get_subgraphs(inputs, self.score_attribute, self.group_attribute, self.score_threshold, self.min_size,
                                         self.max_size)
 
-    def _transform_single(self, graph, score_attribute='importance', group='class', apply_name_estimation=False):
+
+    def transform(self,graphs):
+        return [self._transform_single(g) for g in graphs]
+
+    def _transform_single(self, graph):
         '''
         Parameters
         ----------
@@ -90,36 +91,38 @@ class GraphToAbstractTransformer(object):
         # def f(n,d): d[score_attribute] = graph.degree(n)
         # node_operation(graph,f)
 
-        tcc = ThresholdedConnectedComponents(attribute=score_attribute, more_than=False, shrink_graphs=True)
+        tcc = ThresholdedConnectedComponents(attribute=self.score_attribute, more_than=False, shrink_graphs=True)
         components = tcc._extract_ccomponents(graph, threshold=self.score_threshold, min_size=self.min_size,
                                               max_size=self.max_size)
 
         nodeset = {n for g in components for n in g.nodes()}
 
         def f(n, d):
-            d[group] = '~' if n in nodeset else '-'
+            d[self.group_attribute] = '~' if n in nodeset else '-'
 
         node_operation(graph, f)
 
         # now we either contract what we have, or additionally rename the contracted nodes according to the group estimator
-        if apply_name_estimation:
-            graph = name_estimation(graph, group, self.layer, graphcopy, self.vectorizer, self.nameestimator,
+
+        graph = name_estimation(graph, self.group_attribute, self.layer, graphcopy, self.vectorizer, self.nameestimator,
                                     components)
-        else:
-            graph = contraction([graph],
-                                contraction_attribute=group,
-                                modifiers=[],
-                                nesting=False, dont_contract_attribute_symbol='-').next()
+        #else:
+        #    graph = contraction([graph],
+        #                        contraction_attribute=group,
+        #                        modifiers=[],
+        #                        nesting=False, dont_contract_attribute_symbol='-').next()
 
         graph = nx.relabel_nodes(graph, dict(
             zip(graph.nodes(), range(maxnodeid + 1, 1 + maxnodeid + graph.number_of_nodes()))), copy=False)
+
+        graph.graph['original']= graphcopy
+        graph.graph['layer']=self.layer
         return graph
 
 
 def name_estimation(graph, group, layer, graphreference, vectorizer, nameestimator, subgraphs):
     if subgraphs:
-        map(cleaner, subgraphs)
-
+        map(remove_eden_annotation, subgraphs)
         data = vectorizer.transform(subgraphs)
         clusterids = nameestimator.predict(data)
 
@@ -144,6 +147,7 @@ def name_estimation(graph, group, layer, graphreference, vectorizer, nameestimat
     return graph
 
 
+
 def standalone_get_subgraphs(inputs, score_attribute, group, threshold, min_size, max_size):
     flatter = Flatten()
     tcc = ThresholdedConnectedComponents(attribute=score_attribute, more_than=False, shrink_graphs=True,
@@ -152,18 +156,6 @@ def standalone_get_subgraphs(inputs, score_attribute, group, threshold, min_size
                                          max_size=max_size)
     for e in flatter.transform(tcc.transform(inputs)):
         yield e
-
-
-def hash_function(vec):
-    return hash(tuple(vec.data + vec.indices))
-
-
-def cleaner(graph):
-    # eden contaminates graphs with all sorts of stuff..
-    weight=lambda x:node_operation(x, lambda n, d: d.pop('weight', None))
-    weight(graph)
-    eden.graph._clean_graph(graph)
-    return graph
