+++ graphlearn/minor/autoabstract/cascade.py
@@ -9,18 +9,20 @@ import sklearn.cluster as cluster
 
 
 class Cascade():
-    def __init__(self, depth=2, decomposer=decompose.MinorDecomposer(), debug=False, multiprocess=False):
+    def __init__(self,  depth=2,
+                        debug=False,
+                        multiprocess=True):
+
         self.depth = depth
-        self.decomposer = decomposer
         self.debug = debug
         self.multiprocess = multiprocess
+
+
 
     def setup_transformers(self):
         self.transformers = []
         for i in range(self.depth):
             transformer = transform.GraphMinorTransformer(
-                # subgraph_cluster= cluster.DBSCAN(),# MiniBatchKMeans(n_clusters=8),
-                # save_graphclusters=True,
                 group_score_threshold=-.5,
                 group_max_size=6,
                 group_min_size=3,
@@ -31,21 +33,17 @@ class Cascade():
             self.transformers.append(transformer)
 
     def fit_transform(self, graphs):
-        # create some transformers
+        # INIT
         self.setup_transformers()
-        graphs = self.write_layer_to_graphs(graphs, 0)
+
         # fitting
         for i in range(self.depth):
-            print 'training transformer # %d' % i
-            # graphs = [minor_to_nested(minor)
-            graphs = [minor
-                      for minor in self.transformers[i].fit_transform(graphs)]
-
             if self.debug:
-                print 'graphs at this level'
-                # draw.graphlearn_layered(graphs[:4], vertex_label='layer')
-                print 'all the clusters'
-                draw.graphlearn_dict(self.transformers[i].graphclusters, title_key='hash_title',size=2, edge_label='label')
+                print 'graphs at level %d' % i
+                draw.graphlearn_layered(graphs[:4])
+                #draw.graphlearn_dict(self.transformers[i]..graphclusters, title_key='hash_title',size=2, edge_label='label')
+
+            graphs = self.transformers[i].fit_transform(graphs)
         return graphs
 
     def fit(self, graphs):
@@ -53,7 +51,6 @@ class Cascade():
         return self
 
     def transform(self, graphs):
-        graphs = list(self.write_layer_to_graphs(graphs, 0))
         for i in range(self.depth):
             graphs = self.transformers[i].transform(graphs)
         return graphs
@@ -62,10 +59,7 @@ class Cascade():
         graphs=self.transform(graphs)
         return map(self.remove_intermediary_layers , graphs )
 
-
     def remove_intermediary_layers(self,graph):
-        # write the contracted set to the master layer
-
         def rabbithole(g, n):
             # wenn base graph dann isses halt n
             if 'original' not in g.graph:
@@ -79,19 +73,9 @@ class Cascade():
 
         for n,d in graph.nodes(data=True):
             d['contracted']= rabbithole(graph,n)
-
         # ok get rid of intermediary things
         supergraph=graph
         while 'original' in graph.graph:
             graph = graph.graph['original']
         supergraph.graph['original']=graph
         return supergraph
-
-
-
-    def write_layer_to_graphs(self, graphs, layerid):
-        # just write the layer id in every node that does not yet have a layer attribute :)
-        for graph in graphs:
-            for n, d in graph.nodes(data=True):
-                d['layer'] = layerid
-            yield graph
